Move Stroke to the dc_bundle crate
