+++ crates/figma_import/src/vector_schema.rs
@@ -20,7 +20,7 @@ use dc_bundle::legacy_definition::modifier::shadow::BoxShadow;
 use dc_bundle::legacy_definition::modifier::transform::AffineTransform;
 use serde::{Deserialize, Serialize};
 
-use crate::toolkit_style::Stroke;
+use dc_bundle::legacy_definition::element::path::Stroke;
 
 // This is a simple vector representation that should reproduce Figma or
 // another design tool's rendering. It simplifies the renderer's task by
